Finished Testing
The CRUD process for Alert Contacts should be complete now. Added some documentation to help with it since there's some ambiguity. I should probably look into making a different class for the Type of contact that a C-UD operation returns, but I'm not sure how I want to tackle that.